requirements.txt
Django>=5.1,<5.3
djangorestframework>=3.15
djangorestframework-simplejwt>=5.4
django-filter>=24.3
django-cors-headers>=4.4
django-environ>=0.11
drf-spectacular>=0.27
psycopg[binary]>=3.2
Pillow>=10.4
django-simple-history>=3.7
openpyxl>=3.1
reportlab>=4.2
